# docs/openapi.yaml
openapi: 3.1.0
info:
  title: piper API
  version: 0.1.0
  description: |
    Project-owned resources are always exposed below `/api/projects/{project_id}`.
    Collection POST operations return `201 Created`, successful mutations with no
    response body return `204 No Content`, JSON field names use snake_case, and
    errors use the shape `{"error":"message"}`.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/projects/{project_id}/runs:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          schema: { type: string }
        - name: pipeline_name
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Runs. Step summaries are included only when include_steps=true.
          headers:
            X-Total-Count: { $ref: "#/components/headers/TotalCount" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RunWithSteps" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [yaml]
              properties:
                yaml: { type: string }
                params: { type: object, additionalProperties: true }
                experiment: { type: string }
                vars: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Created run id.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string }
  /api/projects/{project_id}/runs/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Run detail.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run: { $ref: "#/components/schemas/Run" }
                  steps:
                    type: array
                    items: { $ref: "#/components/schemas/Step" }
...
  /api/projects/{project_id}/schedules/{id}/runs:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      parameters:
        - $ref: "#/components/parameters/ScheduleID"
      responses:
        "200":
          description: Runs created by the schedule.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Run" }
  /api/projects/{project_id}/pipeline-templates/{id}/run:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    post:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                params: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Started run.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: string }
...
    RunID:
      name: id
      in: path
      required: true
      schema: { type: string }
    ScheduleID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      description: Page size. When set, the response carries `X-Total-Count`.
      schema: { type: integer, minimum: 1 }
    Offset:
      name: offset
      in: query
      description: Rows to skip; only applies together with `limit`.
      schema: { type: integer, minimum: 0 }
  headers:
    TotalCount:
      description: Total rows matching the filter, ignoring limit/offset. Sent whenever `limit` is.
      schema: { type: integer }
  schemas:
...
    Run:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        schedule_id: { type: string }
        experiment: { type: string }
        pipeline_name: { type: string }
        pipeline_version: { type: integer, description: Template version the run was started from, when it came from a pipeline template. }
        status: { type: string, enum: [scheduled, running, success, failed, canceled] }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        scheduled_at: { type: string, format: date-time }
        pipeline_yaml: { type: string }
        params_json: { type: string }
        created_by: { type: string }
    RunWithSteps:
      allOf:
        - $ref: "#/components/schemas/Run"
        - type: object
          properties:
            steps:
              type: array
              items: { $ref: "#/components/schemas/Step" }
    Step:
      type: object
      properties:
        run_id: { type: string }
        step_name: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        attempts: { type: integer }
        error: { type: string }
...